Chef Shift Audits
At some point, we found that we lacking chef availability in certain days and certain hours.
As a result, WoodSpoon stated to compensate chefs for availability in weekend and rush hours.
This query made to measure chef inactivity time during and before shift ,to see if chef was available to an in advance orders, to compensate chefs for their net availability duration and not gross.
This query saved WoodSpoon money because we could measure the precise amount of availability time instead of gross duration of shift time.