Diners Rank By Month
Showing monthly rank for every diner.
New Diner - a diner that created a user in the past 3 months and placed an order.
Regular Diner - a diner that already made its first order in previous months, and has 1-3 orders in current month.
Super Diner - a diner that already made its first order in previous months, and has 4 orders or more in current month.
Absent 1/2/3 months - a new user OR a diner that didn't order for 1/2/3 months.
Churned Diner - a new user OR a diner that didn't order for 4 months or more.
Revived Diner - a diner that ordered in current month and got back from being a churned diner.
Ranks Cannot overlap in the same month, a new diner wont be also a regular / super diner.